Allow to us `--target` flag with Private Instances
Using --target flag only works with instance with public DNS record due to the login in the `setMultiTarget()` function.
There is an if condition that checks for the `domain`. Private Instances does not have a domain yet you can connect to them using their instance-id.

I removed this check, and modified the `findDomainByInstanceId()` to return a descriptive error if instance was not found by its ID to accomidate for the removed condition check.

Now I can use `--target` for Private Instances, and I can happily use SSM in my CI/CD <3

Signed-off-by: Sherif Abdel-Naby <sherifabdlnaby@gmail.com>